+++ src/Config.php
@@ -1,0 +1,59 @@
+<?php
+
+namespace MetaRush\DataMapper;
+
+class Config
+{
+    private $adapter = 'AtlasQuery';
+    private $dsn;
+    private $dbUser;
+    private $dbPass;
+
+    public function getAdapter(): string
+    {
+        return $this->adapter;
+    }
+
+    public function setAdapter(string $adapter)
+    {
+        $this->adapter = $adapter;
+
+        return $this;
+    }
+
+    public function getDsn(): string
+    {
+        return $this->dsn;
+    }
+
+    public function setDsn(string $dsn)
+    {
+        $this->dsn = $dsn;
+
+        return $this;
+    }
+
+    public function getDbUser(): ?string
+    {
+        return $this->dbUser;
+    }
+
+    public function setDbUser(?string $dbUser)
+    {
+        $this->dbUser = $dbUser;
+
+        return $this;
+    }
+
+    public function getDbPass(): ?string
+    {
+        return $this->dbPass;
+    }
+
+    public function setDbPass(?string $dbPass)
+    {
+        $this->dbPass = $dbPass;
+
+        return $this;
+    }
+}