+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.3
 
 - Allow null dbUser and dbPass
+- Add separate Config class
 
 ## 1.0.2